Move settings UI handling to separate file
